split display row into it's own component
